+++ CHANGELOG.md
@@ -26,6 +26,12 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/jira-actions/compare/release-3.12.0...master
 
+### Added
+- Added top navigation bar handler to API - `com.atlassian.performance.tools.jiraactions.api.page.TopNav`
+
+### Fixed
+- Issue creation time will no longer include the dashboard loading time
+
 ## [3.12.0] - 2020-04-15
 [3.12.0]: https://github.com/atlassian/jira-actions/compare/release-3.11.0...release-3.12.0